exception #2

+++ src/player/player.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common'
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common'
 import { InjectModel } from '@nestjs/mongoose'
 import { DeleteResult } from 'mongodb'
 import { Model } from 'mongoose'
@@ -9,19 +9,24 @@ import { Player } from './player.interface'
 export class PlayerService {
   constructor(@InjectModel('Player') private readonly player: Model<Player>) {}
 
-  async save(player: SavePlayerDto): Promise<Player> {
-    const { _id } = player
-    const find = await this.find({ _id })
+  // async save(player: SavePlayerDto): Promise<Player> {
+  //   const { _id } = player
+  //   const find = await this.find({ _id })
 
-    if (find) {
-      // throw new BadRequestException(`Player\`s email(${player.email}) already registered.`)
-      return await this.update(player)
-    } else {
-      return await this.insert(player)
-    }
-  }
+  //   if (find) {
+  //     // throw new BadRequestException(`Player\`s email(${player.email}) already registered.`)
+  //     return await this.update(player)
+  //   } else {
+  //     return await this.insert(player)
+  //   }
+  // }
 
   async remove(_id?: string): Promise<DeleteResult> {
+    const notFound = !(await this.player.findOne({ _id }).exec())
+    if (notFound) {
+      throw new NotFoundException(`Player id: ${_id} not found`)
+    }
+
     return await this.player.deleteOne({ _id }).exec()
   }
 
@@ -34,22 +39,31 @@ export class PlayerService {
   }
 
   async findById(_id?: string): Promise<Player> {
+    const notFound = !(await this.player.findOne({ _id }).exec())
+    if (notFound) {
+      throw new NotFoundException(`Player id: ${_id} not found`)
+    }
+
     return await this.player.findOne({ _id }).exec()
   }
 
-  async update({ _id, ...player }: SavePlayerDto): Promise<Player> {
-    const notfound = !this.findById(_id)
-
-    if (notfound) {
-      throw new BadRequestException(`Player\`s email already registered.`)
+  async update({ _id, ...player }: SavePlayerDto): Promise<void> {
+    const notFound = !(await this.player.findOne({ _id }).exec())
+    if (notFound) {
+      throw new NotFoundException(`Player id: ${_id} not found`)
     }
 
-    return await this.player
-      .findOneAndUpdate({ _id }, { $set: { ...player } }, { upsert: true })
-      .exec()
+    await this.player.findOneAndUpdate({ _id }, { $set: { ...player } }, { upsert: true }).exec()
   }
 
-  private async insert(player: SavePlayerDto): Promise<Player> {
+  async insert(player: SavePlayerDto): Promise<Player> {
+    const { email } = player
+    const find = await this.player.findOne({ email }).exec()
+
+    if (find) {
+      throw new BadRequestException(`Player's email ${email} already registered.`)
+    }
+
     const created = new this.player(player)
     return await created.save()
   }
